pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jobtool"
version = "1.0.0"
description = "UK job application CLI tool with ATS-compliant CV generation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "TOP G LABS LTD"}
]
keywords = ["job", "cv", "resume", "ats", "cli", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "playwright>=1.40.0",
    "python-docx>=1.1.0",
    "anthropic>=0.18.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
jobtool = "jobtool.cli:app"

[project.urls]
Homepage = "https://github.com/topglabs/jobtool"

[tool.setuptools.packages.find]
where = ["."]
include = ["jobtool*"]
